updated SPI driver

+++ STM32_2nd_batch/Driver_Dev_SPI_GPIO_I2C_UART/Driver_Dev/Drivers/Inc/stm32f446xx_spi_driver.h
@@ -105,7 +105,7 @@ void SPI_PericClockControl(SPI_RegDef_t *pSPIx, uint8_t EnorDi);
 void SPI_Init(SPI_Handle_t *pSPIHandle);
 void SPI_DeInit(SPI_RegDef_t *pSPIx);
 
-void SPI_SendData(SPI_RegDef_t *pSPIx, uint8_t *pTxBuffer, uint8_t Len);
+void SPI_SendData(SPI_RegDef_t *pSPIx,uint8_t *pTxBuffer, uint32_t Len);
 void SPI_ReceiveData(SPI_RegDef_t *pSPIx, uint8_t *pRxBuffer, uint8_t Len);
